feat: ErrorBoundary에 대한 타입폴더로 분리하여 관리
