+++ frontend/src/components/helper/CommonErrorBoundary.tsx
@@ -1,14 +1,9 @@
 import { Component } from 'react';
 
-import CustomError from '@/components/helper/CustomError';
-
-export interface ErrorBoundaryProps {
-	children?: React.ReactNode;
-}
-
-export interface ErrorBoundaryState {
-	error: CustomError | null;
-}
+import {
+	ErrorBoundaryProps,
+	ErrorBoundaryState,
+} from '@/components/helper/types/ErrorBoundary.type';
 
 class CommonErrorBoundary<P> extends Component<P & ErrorBoundaryProps, ErrorBoundaryState> {
 	constructor(props: P & ErrorBoundaryProps) {
